+++ kitchenpos-menu/src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.exception.ErrorMessage;
+import kitchenpos.common.ErrorMessage;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
