+++ kitchenpos-menu/src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.exception.ErrorMessage;
+import kitchenpos.common.ErrorMessage;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
